requirements.txt
identity[quart]>=0.9,<0.10
quart-session[redis]
httpx
python-dotenv<0.22
